% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{lowercase_clean_dataframe}
\alias{lowercase_clean_dataframe}
\title{Cleans a dataframe, removing '*', replacing spaces with '_' and changing all text to lower case.}
\usage{
lowercase_clean_dataframe(df)
}
\arguments{
\item{df}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Cleans a dataframe, removing '*', replacing spaces with '_' and changing all text to lower case.
}
\examples{
lower_clean_dataframe(df)
}
